refactor(code-refactoring): implement feedback
- verify token for every request
- check for 500 error in every request
- account for 500 error on the client side

[Delivers]
