+++ client/actions/UserActions.js
@@ -1,9 +1,9 @@
 import axios from 'axios';
 import { browserHistory } from 'react-router';
 
-import authorization from '../utils/authorization';
+import Auth from '../utils/Auth';
 import * as actionTypes from './actionTypes';
-import displayError from '../utils/errorDisplay';
+import toastMessage from '../utils/toastMessage';
 /**
  * @class UserActions
  */
@@ -35,7 +35,7 @@ export default class UserActions {
           dispatch({ type: actionTypes.REGISTRATION_SUCCESSFUL,
             payload: response.data });
           browserHistory.push('/signin');
-          return displayError('Registration successful. Kindly login here');
+          return toastMessage('Registration successful. Kindly login here');
         })
         .catch((err) => {
           if (err.response.status === 500) {
@@ -45,7 +45,7 @@ export default class UserActions {
             dispatch({ type: actionTypes.REGISTRATION_UNSUCCESSFUL,
               payload: err.response.data });
           }
-          return displayError(err.response.data.message);
+          return toastMessage(err.response.data.message);
         });
     };
   }
@@ -71,11 +71,11 @@ export default class UserActions {
       })
         .then((response) => {
           dispatch({ type: 'LOGIN_SUCCESSFUL', payload: response.data });
-          authorization(response.data.user.token);
+          Auth.setToken(response.data.user.token);
           localStorage.setItem('user',
             JSON.stringify(response.data.user));
           browserHistory.push('/dashboard');
-          return displayError('You are now logged in');
+          return toastMessage('You are now logged in');
         })
         .catch((err) => {
           if (err.response.status === 500) {
@@ -85,7 +85,7 @@ export default class UserActions {
             dispatch({ type: 'LOGIN_UNSUCCESSFUL',
               payload: err.response.data });
           }
-          return displayError(err.response.data.message);
+          return toastMessage(err.response.data.message);
         });
     };
   }
@@ -95,7 +95,7 @@ export default class UserActions {
   *
   * @static
   *
-  * @param {Integer} groupId The id of the group to get members for
+  * @param {Number} groupId The id of the group to get members for
   * @param {String} token the login token
   *
   * @returns {Object} dispatch object
@@ -110,11 +110,12 @@ export default class UserActions {
       .catch((err) => {
         if (err.response.status === 500) {
           dispatch({ type: 'GET_MEMBERS_REJECTED',
-            payload: 'Sorry, an unexpected error occurred.' });
+            payload: err.response.data.message });
         } else {
           dispatch({ type: 'GET_MEMBERS_FAILED',
             payload: err.response.data.message });
         }
+        return toastMessage(err.response.data.message);
       });
   }
 
@@ -123,7 +124,7 @@ export default class UserActions {
   *
   * @static
   *
-  * @param {Integer} groupId The id of the group to add members to
+  * @param {Number} groupId The id of the group to add members to
   * @param {String} userId The id's of the users to add
   * @param {String} token The JWToken to access the endpoint
   *
@@ -138,17 +139,17 @@ export default class UserActions {
     })
       .then((response) => {
         dispatch({ type: 'MEMBERS_ADDED', payload: response.data.message });
-        return displayError('Users added');
+        return toastMessage('Users added');
       })
       .catch((err) => {
         if (err.response.status === 500) {
           dispatch({ type: 'ADD_MEMBERS_REJECTED',
-            payload: 'Sorry, an unexpected error occurred.' });
+            payload: err.response.data.message });
         } else {
           dispatch({ type: 'ADD_MEMBERS_FAILED',
             payload: err.response.data.message });
         }
-        return displayError(err.response.data.message);
+        return toastMessage(err.response.data.message);
       });
   }
 
@@ -174,7 +175,7 @@ export default class UserActions {
         .then((response) => {
           dispatch({ type: 'RESET_PASSWORD_SUCCESSFUL',
             payload: response.data.message });
-          return displayError(response.data.message);
+          return toastMessage(response.data.message);
         })
         .catch((err) => {
           if (err.response.status === 500) {
@@ -184,7 +185,7 @@ export default class UserActions {
             dispatch({ type: 'RESET_PASSWORD_UNSUCCESSFUL',
               payload: err.response.data.message });
           }
-          return displayError(err.response.data.message);
+          return toastMessage(err.response.data.message);
         });
     };
   }
@@ -213,11 +214,12 @@ export default class UserActions {
         .catch((err) => {
           if (err.response.status === 500) {
             dispatch({ type: 'VERIFY_PASSWORD_REJECTED',
-              payload: 'Sorry, an unexpected error occurred.' });
+              payload: err.response.data.message });
           } else {
             dispatch({ type: 'VERIFY_PASSWORD_UNSUCCESSFUL',
               payload: err.response.data.message });
           }
+          return toastMessage(err.response.data.message);
         });
     };
   }
@@ -253,6 +255,7 @@ export default class UserActions {
             dispatch({ type: 'REGISTER_GOOGLE_USER_UNSUCCESSFUL',
               payload: err.response.data });
           }
+          return toastMessage(err.response.data.message);
         });
     };
   }
@@ -292,6 +295,7 @@ export default class UserActions {
             dispatch({ type: 'SEND_EMAIL_NOTIFICATION_UNSUCCESSFUL',
               payload: err.response.data.message });
           }
+          return toastMessage(err.response.data.message);
         });
     };
   }
@@ -321,11 +325,12 @@ export default class UserActions {
         .catch((err) => {
           if (err.response.status === 500) {
             dispatch({ type: 'SEND_SMS_NOTIFICATION_REJECTED',
-              payload: 'Sorry, an unexpected error occurred.' });
+              payload: err.response.data.message });
           } else {
             dispatch({ type: 'SEND_SMS_NOTIFICATION_UNSUCCESSFUL',
               payload: err.response.data.message });
           }
+          return toastMessage(err.response.data.message);
         });
     };
   }
@@ -335,8 +340,8 @@ export default class UserActions {
   *
   * @static
   *
-  * @param {Integer} messageId The id of the message to get readers for
-  * @param {Integer} groupId The id of the group the message belongs
+  * @param {Number} messageId The id of the message to get readers for
+  * @param {Number} groupId The id of the group the message belongs
   *
   * @returns {Object} dispatch object
   *
@@ -365,7 +370,7 @@ export default class UserActions {
   *
   * @static
   *
-  * @param {integer} userId id of the user that owns the notification
+  * @param {Number} userId id of the user that owns the notification
   * @param {String} groupName name of the group  
   * @param {String} message message posted
   * @param {String} postedby the poster
@@ -403,7 +408,7 @@ export default class UserActions {
   *
   * @static
   *
-  * @param {Integer} userId The id of the user to get notification for
+  * @param {Number} userId The id of the user to get notification for
   *
   * @returns {Object} dispatch object
   *
@@ -422,6 +427,7 @@ export default class UserActions {
           } else {
             dispatch({ type: 'GET_NOTIFICATION_UNSUCCESSFUL' });
           }
+          return toastMessage(err.response.data.message);
         });
     };
   }
@@ -431,7 +437,7 @@ export default class UserActions {
   *
   * @static
   *
-  * @param {Integer} userId The id of the user to delete notification
+  * @param {Number} userId The id of the user to delete notification
   *
   * @returns {Object} dispatch object
   *
@@ -450,6 +456,7 @@ export default class UserActions {
           } else {
             dispatch({ type: 'DELETE_NOTIFICATION_UNSUCCESSFUL' });
           }
+          return toastMessage(err.response.data.message);
         });
     };
   }
@@ -460,7 +467,7 @@ export default class UserActions {
   *
   * @static
   *
-  * @param {Array} currentMembers The id of the user to delete notification
+  * @param {Array} currentMembers The current members of the group
   *
   * @returns {Object} dispatch object
   *
@@ -468,10 +475,8 @@ export default class UserActions {
   */
   static getSearchedUsers(currentMembers) {
     return (dispatch) => {
-      return axios.get('/api/v1/users', {
-        headers: {
-          currentmembers: currentMembers
-        }
+      return axios.post('/api/v1/users', {
+        currentMembers
       })
         .then((response) => {
           dispatch({ type: 'GET_SEARCHED_USERS_SUCCESSFUL',
@@ -483,6 +488,7 @@ export default class UserActions {
           } else {
             dispatch({ type: 'GET_SEARCHED_USERS_UNSUCCESSFUL' });
           }
+          return toastMessage(err.response.data.message);
         });
     };
   }
