--- client/utils/authorization.js
@@ -1,9 +1,0 @@
-import axios from 'axios';
-
-const setToken = (token) => {
-  if (token) {
-    axios.defaults.headers.common.token = token;
-  }
-};
-
-export default setToken;
