+++ client/utils/toastMessage.js
@@ -1,8 +1,8 @@
 
-const displayError = (message) => {
+const toastMessage = (message) => {
   Materialize.toast(
     message, 2000,
     'red darken-4 white-text rounded');
 };
 
-export default displayError;
+export default toastMessage;
