+++ server/routes/user.js
@@ -1,18 +1,17 @@
 import express from 'express';
 
 import UserControllers from '../controllers/UserControllers';
-import GroupControllers from '../controllers/GroupControllers';
-import authenticate from '../helpers/Auth';
-import validate from '../helpers/Validation';
+import Auth from '../helpers/Auth';
+import Validate from '../helpers/Validate';
 
 const user = express.Router();
 
 user.post('/api/v1/user/signup',
-  validate.checkPassword,
+  Validate.checkPassword,
   UserControllers.signUp
 );
 user.post('/api/v1/user/signin',
-  validate.checkUser,
+  Validate.checkUser,
   UserControllers.signIn
 );
 user.post('/api/v1/user/reset-password',
@@ -22,29 +21,34 @@ user.post('/api/v1/user/email/verify',
   UserControllers.verifyPassword
 );
 user.post('/api/v1/user/email',
+  Auth.verifyToken,
   UserControllers.mailNotification
 );
 user.post('/api/v1/user/sms',
+  Auth.verifyToken,
   UserControllers.smsNotification
 );
 user.post('/api/v1/user/signup/google',
   UserControllers.registerGoogleUser
 );
-user.get('/api/v1/users',
-  authenticate.verifyToken,
+user.post('/api/v1/users',
+  Auth.verifyToken,
   UserControllers.getAvailableUsers
 );
 user.post('/api/v1/user/notification',
+  Auth.verifyToken,
   UserControllers.saveNotification
 );
 user.get('/api/v1/user/:userId/notification',
+  Auth.verifyToken,
   UserControllers.getNotifications
 );
 user.get('/api/v1/user/groups',
-  authenticate.verifyToken,
-  GroupControllers.getUserGroups
+  Auth.verifyToken,
+  UserControllers.getUserGroups
 );
 user.delete('/api/v1/user/:userId/notification',
+  Auth.verifyToken,
   UserControllers.deleteNotification
 );
