+++ server/server.js
@@ -21,7 +21,7 @@ app.use((req, res, next) => {
   res.header('Access-Control-Allow-Origin', '*');
   res.header('Access-Control-Allow-Methods', 'GET,HEAD,OPTIONS,POST,PUT');
   res.header('Access-Control-Allow-Headers',
-    'Access-Control-Allow-Headers, Origin,Accept, X-Requested-With, Content-Type, token, currentmembers, mailToken');
+    'Access-Control-Allow-Headers, Origin,Accept, X-Requested-With, Content-Type, token, mailToken');
   next();
 });
